+++ lib/db/student.py
@@ -16,23 +16,3 @@ class Student(Base):
     contact_by_phone = Column(Boolean)
 
     reports = relationship("Report", back_populates="student")
-
-class Report(Base): 
-
-    __tablename__ = "reports"
-
-    id = Column(Integer, primary_key=True)
-    grade = Column(Integer)
-    days_enrolled = Column(Float)
-    days_present = Column(Float)
-    days_excused = Column(Float)
-    days_not_excused = Column(Float) 
-
-    student_id = Column(Integer, ForeignKey('student.id')) 
-
-    student = relationship("Student", back_populates="reports")
-    
-    #def render(self, _template):
-    #    tm = jinja.Template(_template)
-    #    return tm.render(student=self)
-
